Fix bug in reset password and add static files.

+++ Data/DbInitializer.cs
@@ -2,15 +2,36 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
 using XueLeMeBackend.Models;
 using XueLeMeBackend.Models.Fragments;
 namespace XueLeMeBackend.Data
 {
     public class DbInitializer
     {
-        public static void Init(XueLeMeContext context)
+        public DbInitializer(ILogger<DbInitializer> logger, IConfiguration configuration)
         {
-            context.Database.EnsureDeleted();
+            Initialized = false; 
+            Logger = logger;
+            Configuration = configuration;
+        }
+        public bool Initialized { get; set; }
+        public ILogger<DbInitializer> Logger { get; }
+        public IConfiguration Configuration { get; }
+
+        public void Init(XueLeMeContext context) 
+        {
+            if (!Initialized) {
+                Initialized = true;
+                DoInit(context);
+            }
+        }
+        public void DoInit(XueLeMeContext context)
+        {
+            if (!Configuration.GetValue<bool>("IsServer")) {
+                context.Database.EnsureDeleted();
+            }
             context.Database.EnsureCreated();
             var mail = "614434935@qq.com";
             var user = new User
@@ -25,6 +46,7 @@ namespace XueLeMeBackend.Data
                 AuthToken = mail,
                 CreatedTime = DateTime.Now
             };
+            user.Authentications.Add(auth);
             context.Users.Add(user);
 
             var mail2 = "jamesnm2015@163.com";
@@ -40,7 +62,7 @@ namespace XueLeMeBackend.Data
                 AuthToken = mail2,
                 CreatedTime = DateTime.Now
             };
-            user.Authentications.Add(auth2);
+            user2.Authentications.Add(auth2);
             context.Users.Add(user2);
             context.SaveChanges();
         }
